Started off with this issue: Google civic returns the same google_civic_election_id for primaries that occur in multiple states on the same day, so we need to persist the state_code most places where we store google_civic_election_id. Added (working) error handling when receiving un-successful API queries. Error response in these error conditions provides text that describes the problem, instead of silently failing.
There was so much inconsistently "repeated" code, that resulted in
silently ignored errors, so I added process_request_from_master().

Reworked sync_data_with_master_dashboard.html so that it was more
more understandable by newbies, and so that errors bubbled up to the
web page (instead of failing silently).

This change requires a DB migration.