Server side #427, Update election/views_admin.py elections_sync_out_view to use the read-only database. In environment_variables-template.json add the block of settings for the read only slave replicated postgres server. In config/local.py add the DATABASES section for the read only server. In controllers.py change the "Election.objects.all()" to "Election.objects.using('readonly').all()" so that it uses the readonly db server.
Each engineer on the project will need to make the changes to their
local environment_variables.json -- Add the block of 7 lines with the
_READONLY vars.  It is safest to just copy all the settings from the
non readonly settings to the readonly.  In that way if you try to make
a request to your local on port 8000, you will get the read/write
database through the read/only settings.

This has been tested with two postgres servers running on two macs.